requirements.txt
alembic==1.8.1
anyio==3.6.2
bcrypt==4.0.1
boto3==1.26.153
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.86.0
greenlet==2.0.0.post0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
Mako==1.2.3
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.5
pyasn1==0.4.8
pydantic==1.10.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
rsa==4.9
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing-extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uvicorn==0.19.0
zipp==3.10.0
